refactor: message 파일명에 'constants' 추가
상수 파일명의 끝에 'constants'를 붙여 파일의 역할을 명확히 나타냄
